Split 'WatchVariable' into 'MemoryDescriptor' and 'NamedVariableCollection.IVariableView'
Variable panels now use 'NamedVariableCollection.IVariableView's to render, which may be independent from any game memory associations. Instead, specific implementations use 'MemoryDescriptor's to forward such data to the panel, whereas other implementations are used to implement computed values.
All instances of 'NamedVariableCollection.IVariableView' are typesafe with regard to the value they hold.

Implementation for fixing addresses has been temporarily removed.

Implementation for locking variable values has been temporarily removed.

+++ STROOP/Tabs/CoinTab.cs
@@ -95,7 +95,7 @@ namespace STROOP.Tabs
 
             List<int> rngIndexes = Enumerable.Range(0, 65114).ToList();
 
-            foreach (int rngIndex in rngIndexes)
+            foreach (ushort rngIndex in rngIndexes)
             {
                 // rng based values
                 ushort rngValue = RngIndexer.GetRngValue(rngIndex);
